Adding torch.hub support
